Test: 서버 데이터 받아오기 테스트 완료, 추후 콘솔오류 픽스 예정
