+++ .stylelintrc.js
@@ -11,7 +11,7 @@ module.exports = {
   rules: {
     'at-rule-no-unknown': null,
     // scss를 사용하기 때문에 css영역에선 null로 처리합니다.
-    // "selector-class-pattern": "^([a-z][a-z0-9]*)(-[a-z0-9]+)*$",
+    'selector-class-pattern': '^([a-z][a-z0-9]*)(-[a-z0-9]+)*$',
     // Team내 컨벤션으로 수정 (현재 kebab-case)
     'keyframes-name-pattern': /^([a-z][a-z0-9]*)(-[a-z0-9]+)*$/,
     // Team내 컨벤션으로 수정 (현재 kebab-case)
