+++ src/pages/Detail/Detail.js
@@ -1,25 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import './Detail.scss';
 import DetailNavigator from './Detailnavigator/DetailNavigator';
-// 35, 43, 87, 116-135
+import TableInner from './TableInner/TableInner';
+import ShareModal from './ShareModal/ShareModal';
 
 const Detail = () => {
   const [itemInfo, setItemInfo] = useState({});
 
-  const {
-    allerge,
-    detail_image_url,
-    contactant,
-    expiration_date,
-    name,
-    origin,
-    packing_type_text,
-    price,
-    weight,
-    short_description,
-    packing_type_id,
-    packing_types,
-  } = DETAIL_DATA;
+  const { allerge, contactant, expiration_date, name, origin, ...others } =
+    itemInfo;
 
   const DETAIL_KEY_KOREAN = {
     contactant: '판매자',
@@ -33,19 +22,17 @@ const Detail = () => {
   const keys = Object.keys(DETAIL_KEY_KOREAN);
 
   useEffect(() => {
-    fetch('http://url:3000/products/${productId}', {
-      //api 주소 맞추기
+    // fetch('http://10.58.52.148:3000/products/${id}', {
+    fetch('', {
       method: 'GET',
-      headers: { key: 'value' },
     })
       .then(res => res.json())
-      .then(result => setItemInfo(result));
-  });
-  console.log(itemInfo); //서버와 통신 시 확인 위함
+      .then(result => {
+        setItemInfo(result.productData[0]);
+      });
+  }, []);
 
-  // const showShareWindow = ()=>();
-
-  function priceComma(price) {
+  const priceComma = price => {
     const reversedPrice = (price + '').split('').reverse();
     let arr = [];
     for (let i = 0; i < reversedPrice.length; i++) {
@@ -55,14 +42,18 @@ const Detail = () => {
       }
     }
     return arr.reverse().join('');
-  }
+  };
 
   return (
     <div className="detailPage">
       <div className="detailPageContainer">
         <article className="detailArticle">
           <div className="itemImage">
-            <img src={detail_image_url} alt={name} className="itemImageTag" />
+            <img
+              src={others.detail_image_url}
+              alt={name}
+              className="itemImageTag"
+            />
           </div>
           <div className="itemDescription">
             <div className="nameOfItem">
@@ -72,12 +63,12 @@ const Detail = () => {
                   <h2 className="itemName">{name}</h2>
                   <button className="share"></button>
                 </div>
-                <p className="brief">{short_description}</p>
+                <p className="brief">{others.short_description}</p>
               </div>
             </div>
 
             <div className="priceContainer">
-              <span className="price">{priceComma(price)}</span>
+              <span className="price">{priceComma(others.price)}</span>
               <span className="won">원</span>
             </div>
             <div className="table">
@@ -97,7 +88,7 @@ const Detail = () => {
                   <TableInner
                     key={index}
                     name={key}
-                    data={DETAIL_DATA} //itemInfo로 변수명 변경!
+                    data={itemInfo}
                     korean={DETAIL_KEY_KOREAN}
                   />
                 );
@@ -107,43 +98,10 @@ const Detail = () => {
           </div>
         </article>
         <DetailNavigator />
+        <div>추후 이미지와 상품설명이 들어가는 자리입니다.</div>
       </div>
     </div>
   );
 };
 
 export default Detail;
-
-const TableInner = props => {
-  const { name, data, korean } = props;
-  return (
-    <dl className="dList">
-      <dt className="dTitle">{korean[name]}</dt>
-      <dd className="dDescription">{data[name]}</dd>
-    </dl>
-  );
-};
-
-//서버와 통신 전 사용하는 Mockdata 입니다.
-
-const DETAIL_DATA = {
-  id: 1,
-  sub_category_id: 2,
-  name: '[KF365] 밤고구마 800g',
-  thumnail_image_url:
-    'https://pixabay.com/get/g00e4af9aa464d5be8b52b11a34cfd3705596eb00ae6fcc74618239c7b4dfa2e4f5c568ade6c3dc5eec4871edf803545089561841af9fe44d6d9e9367126a0fe8066e329cfa3f2361ca945faa8fe13f60_640.jpg',
-  short_description: '포근하고 고소한 고구마(1봉/800g)',
-  contactant: '컬리',
-  packing_type_id: 1,
-  weight: '800g',
-  origin: '국내산',
-  allerge: '-없음',
-  expiration_date:
-    '농산물로 별도의 유통기한은 없으나 가급적 빨리 드시기 바랍니다.',
-  price: '1113900',
-  detail_image_url:
-    'https://images.unsplash.com/photo-1573865526739-10659fec78a5?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=830&q=80',
-  created_at: '2022-10-20T00:23:09.000Z',
-  updated_at: '2022-10-20T00:23:09.000Z',
-  packing_types: '냉장',
-};
