+++ src/pages/Detail/ShareModal/ShareModal.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ShareModal = () => {
+  return <div></div>;
+};
+
+export default ShareModal;
